Refactor & expand test
New tests for weakBox/weakDiamond
