+++ CIV.Test/CIV.Test.csproj
@@ -23,4 +23,8 @@
     <ProjectReference Include="..\CIV.Hml\CIV.Hml.csproj" />
     <ProjectReference Include="..\CIV.Common\CIV.Common.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Ccs\" />
+    <Folder Include="Hml\" />
+  </ItemGroup>
 </Project>
